otto-atlas-web/ml_pipeline/requirements.txt
fastapi
uvicorn
python-multipart
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastai
requests
psycopg2-binary
